separated code to check which app it is in concern
