+++ app/controllers/application_controller.rb
@@ -1,6 +1,7 @@
 class ApplicationController < ActionController::API
   include CanCan::ControllerAdditions
   include TokenValidatable
+  include AppMatcher
 
   check_authorization
 
@@ -8,32 +9,8 @@ class ApplicationController < ActionController::API
   before_action :set_locale, :current_user
   helper_method :current_user
 
-  def is_admin_app
-    current_user.try(:staff?) && app_name == ADMIN_APP
-  end
-
-  def is_stockit_request
-    app_name == STOCKIT_APP
-  end
-
-  def is_stock_app
-    app_name == STOCK_APP
-  end
-
-  def is_browse_app
-    app_name == BROWSE_APP
-  end
 
   protected
-
-  def app_name
-    request.headers['X-GOODCITY-APP-NAME'] || meta_info["appName"]
-  end
-
-  def meta_info
-    meta = request.headers["X-META"].try(:split, "|")
-    meta ? Hash[*meta.flat_map{|a| a.split(":")}] : {}
-  end
 
   def app_version
     request.headers['X-GOODCITY-APP-VERSION']
